WIP: default transport task type set to Repair
In our system, I must track my changes in Repair tasks. I have no idea
why and I have no power to change it.

The function TRINT_ORDER_CHOICE looks for 'Development/Correction' and
'Unclassified' tasks by default.

Without this patch, it is just not possible to select a transport task
in the system. Because unclassified tasks cannot be classified as
Repairs and if I classify a task manually the function ignores it.

With this patch, the function no longer ignores manually classified
tasks.

I need to find the proper place for this configuration.

Related #1446.