CUA: make error message verbose enough for action
I had the same problem as described in #1899 and displaying the error
code didn't help me much.

I wasn't even able to pull new development version of abapGit because
the function module was reporting the error for zabapgit.

I temporarily removed the function module call, upgraded abapgit,
and the issue magically disappeared with abapGit 1.76.0.

Relates to #1899