ui: add a workaround for JavaGUI dyn form bug
JavaGUI does not pass values of options from forms generated via
JavaScript.

This commit adds enhances the base page class to render stub forms
(empty form elements), so the page's JavaScript does not need to create
the form element on its own.

It is an optional behavior and can be enabled by redefining the method
get_events which should return table of required events and the base
page object will add the empty form elements for them. Every table line
will be converted to a form with id = "form_" + the event's name.

On the action submit SAP event in the common JavaScript, the method
checks if there is the stub form and if so, the method uses it instead
of creating the dynamic one. Otherwise, it follows the old approach that
creates the form at run-time.

--- v2

Use 'null' instead of 'undefined' to check the form was not found in
the JavaScript code.

---

Closes #1866