oo_serilizer: try to prevent Pretty printer from staring GUI
As of now, we have no problem in background mode but we still
reaches the maximum number of GUI sessions in foreground.

Without this commit the function module GUI_IS_AVAILABLE is being
executed.

I tried to create a repo with 8 DDIC element and threads were working.
I tried to create a repo with 5 classes and the gui limit was reached.