repo: add folder logic and ign. sub.pkg to the new repo popup
This patch adds the parameters Folder logic and Ignore subpackages to
the new online repo popup to help us deal with old package which do not
follow the prefix approach for subpackages and to allow us to version
superpackages (which are shared by many developers) in separate
repositories.

Unfortunately, I am not sure how to implement this without touching
interfaces. At least, I tried to keep the interface APIs compatible with
the old code.

Closes #2395